+++ epicore_utils/modules/compute_cores.py
@@ -154,7 +154,6 @@ def gen_landscape(protein_df: pd.DataFrame, mod_pattern: str) -> pd.DataFrame:
     protein_df['whole_epitopes_all'] = [[] for _ in range(len(protein_df))]
 
     for r, row in protein_df.iterrows():
-        #seen_pep={}
         for group, [pep_group_start, pep_group_end] in enumerate(zip(row['grouped_peptides_start'], row['grouped_peptides_end'])):
             seen_pep={}
             start_idx = pep_group_start[0]
@@ -173,13 +172,18 @@ def gen_landscape(protein_df: pd.DataFrame, mod_pattern: str) -> pd.DataFrame:
                     pattern = re.escape(mod_pattern.split(',')[0]) + r'.*?' + re.escape(mod_pattern.split(',')[1])
                     current_seq = re.sub(pattern,"",current_seq)
 
+                # check if peptide is repetitive
+                match = re.search(r'^(.+).*\1$', current_seq)
+
+                # position seen before
                 if str(pep_start) + '-' + str(pep_end) in seen_pep:
-                    seen_seq = seen_pep[str(pep_start) + '-' + str(pep_end)]
-                    check_seen = [(seq == current_seq) for seq in seen_seq]
-                    # current_seq not in list for repetitive regions (dont increase landscape value for repetitive regions)
                     seen_pep[str(pep_start) + '-' + str(pep_end)].append(current_seq)
+                    # increase landscape for non repetitive peptide
+                    if not match:
+                        for pos in range(pep_start, pep_end+1):
+                            group_landscape[pos-start_idx] += 1
                 else: 
-                    # if sequence and position not seen before add the sequence to the landscape
+                   # if position not seen before add the sequence to the landscape
                     seen_pep[str(pep_start) + '-' + str(pep_end)] = [current_seq]
                     for pos in range(pep_start, pep_end+1):
                         group_landscape[pos-start_idx] += 1